# optimizations to statsd plugin