- The buffer was being held longer than required. Releasing it immediately
after its used now.
- Replaced strings.Prefix check with simply checking the first index. Same
with strings.Contains

Microbenchmarks show good improvements:
```go
func BenchmarkPrefix(b *testing.B) {
  str := "-somerandomstring"
  for n := 0; n < b.N; n++ {
    // strings.HasPrefix(str, "-")
    _ = str[0] == '-'
  }
}
```
```
name      old time/op    new time/op    delta
Prefix-4    5.07ns ± 0%    0.40ns ±13%  -92.14%  (p=0.008 n=5+5)

name      old alloc/op   new alloc/op   delta
Prefix-4     0.00B          0.00B          ~     (all equal)

name      old allocs/op  new allocs/op  delta
Prefix-4      0.00           0.00          ~     (all equal)
```
- Optimized parseKeyValue function to replaced strings.Split() check
with a strings.IndexByte() and then using the strings from the index.

Microbenchmarks show good improvements:
```go
func BenchmarkSplitEqual(b *testing.B) {
  str := "key=value"
  for n := 0; n < b.N; n++ {
    // splitted := strings.Split(str, "=")
    // _ = splitted[0]
    // _ = splitted[1]
    ind := strings.IndexByte(str, '=')
    _ = str[:ind]
    _ = str[ind+1:]
  }
}
```
```
name          old time/op    new time/op    delta
SplitEqual-4     112ns ± 1%       5ns ± 0%   -95.92%  (p=0.000 n=5+4)

name          old alloc/op   new alloc/op   delta
SplitEqual-4     32.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)

name          old allocs/op  new allocs/op  delta
SplitEqual-4      1.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
```

### Required for all PRs:

- [x] Signed [CLA](https://influxdata.com/community/cla/).
~- [ ] Associated README.md updated.~
- [x] Has appropriate unit tests.