@agnivade 2017-10-01
I am unable to see the logs of the circle ci failure. The spinner just keeps spinning.