@danielnelson 2017-10-03
The spinner is probably a temporary issue with circle, but I think you will need to rebase for the tests to pass due to an issue on the master branch that has been fixed.