@danielnelson 2017-10-03
What do the larger benchmarks show?