@agnivade 2017-10-03
Rebased latest master and pushed.

Larger benchmarks don't show any noticeable improvements. Most probably because the test runs several order of magnitude larger than the difference the improvements make.

```
name   old time/op    new time/op    delta
UDP-4     873ms ± 9%     850ms ± 5%   ~     (p=0.310 n=5+5)

name   old alloc/op   new alloc/op   delta
UDP-4     168MB ± 8%     165MB ± 6%   ~     (p=0.841 n=5+5)

name   old allocs/op  new allocs/op  delta
UDP-4     3.46M ± 9%     3.39M ± 7%   ~     (p=0.690 n=5+5)
```

Note: You can indeed see slight improvements. But because of the variance, the overall delta nullifies itself.