@danielnelson 2017-10-03
I don't want to bother with minor tweaks unless there is a noticeable improvement overall, since every change comes with the risk of introducing bugs.  I wonder if moving to a structure similar to socket_listener would be helpful?  @phemmer do you know how socket_listener compared to tcp/udp_listener?