@phemmer 2017-10-03
Performance wise? No, I never investigated.