@agnivade 2017-10-05
> I don't want to bother with minor tweaks unless there is a noticeable improvement overall, since every change comes with the risk of introducing bugs.

Aren't the unit tests written for this purpose ?

Bugs can and will always come. IMO that should not stop improvements to the code. Yes, the risk increases proportionally depending on the type of change. In this case, it's a couple of lines change which improves some string comparisons. We should add more tests and increase coverage if there are fears of introducing bugs.

At the end of the day, its an open source project. Feel free to close the PR.