@danielnelson 2017-10-05
> Aren't the unit tests written for this purpose ?

Yes, but realistically they are not and will never be sufficient to completely avoid bugs.  Also, many performance enhancements come at the expense of code readability, so I just want to be sure we are going to see a benefit.  Hope you understand in this case.

That said, I am going to apply the bufPool change from this PR.